requirements.txt
asgiref==3.2.3
Django==3.0
pytz==2019.3
sqlparse==0.3.0
lxml==4.4.0
requests==2.22.0
beautifulsoup4==4.8.0
django-cron==0.5.1
djangorestframework==3.10.2
